+++ playground/composables/vanilla.mjs
@@ -1,0 +1,5 @@
+export const vanillaAMJS = 1
+
+export function vanillaBMJS() {
+  return 'vanillaB'
+}