+++ test/cases/output/acorn/switch.js
@@ -1,3 +1,4 @@
+import { Bar } from 'foobar';
 function foo(mode) {
   switch (mode) {
     case Foo: