+++ test/cases/output/default/import-all.js
@@ -1,0 +1,2 @@
+import * as THREE from 'three';
+const scene = new THREE.Scene()