+++ test/dedupe.test.ts
@@ -33,7 +33,8 @@ describe('dedupeImports', () => {
       ]
     `)
 
-    expect(warnMsg).toMatchInlineSnapshot('"Duplicated imports \\"foo\\", the one from \\"module1\\" has been ignored and \\"module2\\" is used"')
+    expect(warnMsg)
+      .toMatchInlineSnapshot(`"Duplicated imports "foo", the one from "module1" has been ignored and "module2" is used"`)
   })
 
   it('respect explit priority', () => {