+++ test/scan-dirs.test.ts
@@ -1,6 +1,6 @@
 import { join, relative } from 'pathe'
 import { describe, expect, it } from 'vitest'
-import { scanDirExports, toImports } from '../src'
+import { scanDirExports, stringifyImports } from '../src'
 
 describe('scan-dirs', () => {
   it('scanDirExports', async () => {
@@ -67,9 +67,19 @@ describe('scan-dirs', () => {
             "name": "vanillaA",
           },
           {
+            "as": "vanillaAMJS",
+            "from": "vanilla.mjs",
+            "name": "vanillaAMJS",
+          },
+          {
             "as": "vanillaB",
             "from": "vanilla.js",
             "name": "vanillaB",
+          },
+          {
+            "as": "vanillaBMJS",
+            "from": "vanilla.mjs",
+            "name": "vanillaBMJS",
           },
           {
             "as": "VanillaInterface",
@@ -150,7 +160,7 @@ describe('scan-dirs', () => {
       ]
     `)
 
-    expect(toImports(importsResult)).toMatchInlineSnapshot(`
+    expect(stringifyImports(importsResult)).toMatchInlineSnapshot(`
       "import { bar, named } from 'nested/bar/index.ts';
       import { myBazFunction } from 'nested/bar/baz.ts';
       import { subFoo } from 'nested/bar/sub/index.ts';"